python/revng/pypeline/pyproject.toml
#
# This file is distributed under the MIT License. See LICENSE.md for details.
#

[project]
name = "pypeline"
dynamic = ["version", "readme"]
dependencies = [
    "click==8.2.1",
    "jsonschema",
    "pyyaml",
    "psuitl",
    "xdg",
    "starlette",
    "uvicorn"
]

[project.scripts]
pype = "revng.pypeline.main:run"

[tool.setuptools.package-data]
revng = ["pypeline/schema.yml"]

[build-system]
requires = ["setuptools", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools.dynamic]
version = {attr = "revng.pypeline.__version__"}
readme = {file = ["README.md"], content-type = "text/markdown"}
